requirements.txt
setuptools>=58.0.0
wheel>=0.37.0
Flask==3.0.3
gunicorn==20.1.0
Flask_Cors==4.0.0
Flask_RESTful==0.3.10
numpy==2.2.0
Requests==2.31.0
pillow==11.1.0
huggingface_hub==0.20.3
python-dotenv== 1.1.0